Checking in some changes to explicitly cast a litteral to a double when used in a == case.
git-svn-id: https://vlab.ncep.noaa.gov/svn/degrib/degrib@970 98767674-7b1f-0410-897a-ff3155737c25

+++ src/degrib/degrib1.c
@@ -1786,7 +1786,7 @@ int ReadGrib1Record (FILE *fp, sChar f_unit, double **Grib_Data,
    meta->pds1.gridID = gridID;
    /* Allow data originating from NCEP to be 6371.2 by default. */
    if ((meta->center == NMC)) {
-      if (meta->gds.majEarth == 6367.47) {
+      if (meta->gds.majEarth == (double) 6367.47) {
          meta->gds.f_sphere = 1;
          meta->gds.majEarth = 6371.2;
          meta->gds.minEarth = 6371.2;
